HypeHunter-main/requirements.txt
# Add your existing dependencies
numpy==1.21.1
pandas==1.3.1

scikit-learn==0.24.2

streamlit==1.10.0
Pillow==8.3.1

# Add NLTK data dependency
nltk==3.6.2
